Fix several problems that arose out of newer versions of doxygen and LaTeX.
Recent doxygen versions emit yet another list of packages at the top
of the LaTeX source code, so the patch trick that is intented to
replace the doxygen table by a standard longtable failed to work.

Recent LaTeX versions are now based on pdfetex, and obviously are very
sensitive to any LaTeX package calls that might cause the generation
of PDF output.  In that case, even when called "latex" (as opposed to
"pdflatex"), or when called with -output-format=dvi, they switch to
generate PDF on the fly -- and then fail to include the PostScript
pictures, as the PDF driver can only include PDF pictures.  Turn off
everything in doxygen.conf by default that would trigger references to
PDF, and only turn it on when actually generating the PDF version of
the docs.

All changes are supposed to be backwards compatible with previous
versions of doxygen and LaTeX.


git-svn-id: https://svn.savannah.nongnu.org/svn/avr-libc/trunk/avr-libc@768 1084037a-409f-4a99-8160-1a4b3577d4c1