Use autotools to determine Python command to run mlib-gen.py.
The shebang in devtools/mlib-gen.py does not work on all
systems, so let AM_PATH_PYTHON determine what Python is.
The only outcome is that @PYTHON@ in gen-avr-lib-tree.sh.in
will be resolved to the actual Python command.

Contrary to the toplevel, devtools/ comes with configure and
aclocal.m4 scripts.  This means when changes are made to
devtools/configure.ac, then autoconf and maybe aclocal must
be run in devtools, and the re-build files have to be added
to git.

	* bootstrap: Run devtools/configure.
	* .gitignore: Don't ignore: devtools/aclocal.m4, devtools/configure.
devtools/
	* configure.ac: New file.
	* configure: New file, (re)generate.
	* aclocal.m4: New file, (re)generate.
	* gen-avr-lib-tree.sh.in: New file from gen-avr-lib-tree.sh.
	(mlib-gen.py): Don't use as a script. Instead, use Python program
	as determined by configure to run it.
	* Makefile.am (EXTRA_DIST): Add: aclocal.m4, configure.ac, configure.
	[gen-avr-lib-tree.sh.in]: Instead of gen-avr-lib-tree.sh.