Use all .h files of $srcdir/include/avr in Makefile.am. Use all .S files of $srcdir/crt1/iosym in Makefile.am.
We currently have the situation that  include/avr/Makefile.am::avr_HEADERS  has
a list of all headers like iom8.h in that folder.  As new headers are added,
each header must also be mentioned in avr_HEADERS.

The patch generates a list of all headers, so that it is enough to drop
some .h file in include/avr and the build machinery will take care of it.

	* configure.ac [SUBST_BASENAME_FILES]: New AC_DEFUN.
	[AVRLIBC_INCLUDE_HFILES, AVRLIBC_IOSYM_SFILES]:
	New SUBST_BASENAME_FILES.
	* include/avr/Makefile.am [avr_HEADERS]: Use @AVRLIBC_INCLUDE_HFILES@
	instead of mentioning all headers in include/avr.
	* crt1/iosym/Makefile.am [EXTRA_DIST]: Use @AVRLIBC_IOSYM_SFILES@
	instead of mentioning all S files in crt1/iosym.