Issue #931: Initialize NVMCTRL_CTRLB.FLMAP for Devices that have it.
https://sourceware.org/PR31124
https://gcc.gnu.org/PR112944
support locating .rodata in program memory for some devices from the
avrxmega2 (AVR64*) and avrxmega4 (AVR128*) families.

The user can chose the 32 KiB flash block which hosts the .rodata section
by means of defining symbol __flmap.  Or they can return to the old layout
with .rodata in RAM by means of compiler option -mrodata-in-ram.

In the rodata-in-flash case, the startup code in .init2 sets bit-field
NVMCTRL_CTRLB.FLMAP to __flmap.  In the rodata-in-ram case, the startup
code leaves NVMCTRL_CTRLB alone.  The decision is taken by a
conditional RJMP, so that the same crt<mcu>.o serves both cases.

The user can also chose to lock the FLMAP bit-field from further changes
by defining symbol __flmap_lock to a non-0 value.

The code sequence in .init2 will only be activated when __AVR_HAVE_FLMAP__
is defined, which is only the case when avr-gcc implements PR112944, and
avr-ld implements PR31124.

	* crt1/gcrt1.S (section .init2) [__AVR_HAVE_FLMAP__]:
	Weakly define __flmap and __flmap_lock.
	Define global symbols __flmap_init_start and __flmap_noinit_start.
	Initialize NVMCTRL_CTRLB.FLMAP to __flmap.
	Initialize NVMCTRL_CTRLB.FLMAPLOCK to __flmap_lock.
	Define global symbols __flmap_lock_mask and __flmap_bpos which
	describe the layout of FLMAP and FLMAPLOCK in NVMCTRL_CTRLB.