Issue #936: Provide symbols for exact memory layout.
The linker can produce more exact diagnostics, when the properties of
the MEMORY regions as described in the linker description files are
matching the memories available on the hardware.

This is currently not the case:  The defaults for the MEMORY regions
pretend vastly more memory than is actually available.

The canonical place to define respective symbols, which are used by the
default linker scripts, is the startup code in gcrt1.S.

	* crt1/gcrt1.S (__TEXT_REGION_ORIGIN__, __TEXT_REGION_LENGTH__)
	(__DATA_REGION_ORIGIN__, __DATA_REGION_LENGTH__)
	(__EEPROM_REGION_LENGTH__): Weakly define if information is available.