Issue #931: Addendum: Weakly define __flmap_value_with_lock.
When a crt*.o is used with a version of Binutils that don't support PR31124,
-- which may happen when the crt*.o is part of a device support package --
then linking will raise undefined references to __flmap_init_label
and __flmap_value_with_lock.

A way around would be to weakly define these symbols in gcrt1.S, however
the default linker scripts for emulations avrxmega2/4_flmap define symbol
__flmap_init_label as PROVIDE, which means that when PR31124 *IS*
supported, then the initialization of FLMAP would be bypassed and the whole
purpose of issue #931 would be nullified.

The problem is that v2.42 is already out, hence a follow-up patch as proposed
in  https://sourceware.org/pipermail/binutils/2024-January/132247.html
won't go into v2.42 but only in v2.43 (or v2.24.1).  Also the compiler
cannot discriminate between PROVIDE and not PROVIDE in an ld script.

	* crt1/gcrt1.S (section .init2) [__AVR_HAVE_FLMAP__]:
	Weakly define __flmap_value_with_lock.
	Explain why we don't weakly define __flmap_init_label,
	at least for now.