Issue #943: Use _VECTOR_SIZE in gcrt1.S if available.
The size of vectab entries follows the size of XJMP, which is currently
defined in macros.inc according to __AVR_HAVE_JMP_CALL__ as defined by
the compiler. This binds correctness of the vector table to how the
compiler assigns multlilibs.
Some of the newer I/O headers define _VECTOR_SIZE as the size in bytes of
one vectab entry.  This is more reliable -- or at least under better control
of AVR-LibC -- than the multilib assignments of the compiler.

	* crt1/gcrt1.S (XJMP): Define according to _VECTOR_SIZE provided
	_VECTOR_SIZE is defined.
	* include/avr/iom808.h (_VECTOR_SIZE): Define to 4.
	* include/avr/iom809.h (_VECTOR_SIZE): Define to 4.
	* include/avr/ioa5795.h (_VECTOR_SIZE): Define to 2.