Issue #945: Generate a list of all .h files for include/avr/Makefile.am
We currently have the situation that  include/avr/Makefile.am::avr_HEADERS  has
a list of all headers like iom8.h in that folder.  As new headers are added,
each header must also be mentioned in avr_HEADERS and must not be forgotten
like it happened with issue #921.

The patch generates a list of all headers, so that it is enough to drop
some .h file in include/avr and the build machinery will take care of it.

Two files that orignate from *.h.in are treated seperately because they don't
fit the *.h pattern:

- version.h is added to avr_HEADERS by hand.

- builtins.h was not in the original avr_HEADERS because it's not part of
the distribution, but is listed in nodist_avr_HEADERS.

	* configure.ac (SUBST_BASENAME_FILES): New AC_DEFUN.
	(AVRLIBC_INCLUDE_HFILES): New from SUBST_BASENAME_FILES.
	* include/avr/Makefile.am (avr_HEADERS): Use @AVRLIBC_INCLUDE_HFILES@
	instead of mentioning all headers in include/avr.