Issue #945: Generate a list of all .S files for crt1/iosym/Makefile.am
Currently crt1/iosym/Makefile.am::EXTRA_DIST has a list of all assembly
files in that folder.  As new files are added, each one must also be
added to EXTRA_DIST.

The patch generates a list of all .S flies, so that it is enough to drop
some .S file in crt1/iosym and the build machinery will take care of it.

	* configure.ac (AVRLIBC_IOSYM_SFILES): New from SUBST_BASENAME_FILES.
	* crt1/iosym/Makefile.am (EXTRA_DIST): Use @AVRLIBC_IOSYM_SFILES@
	instead of mentioning all assembly files in crt1/iosym.