Issue #931: Add module to lib<mcu>.a to init bit-field NVMCTRL_CTRLB.FLMAP
When .rodata is located in flash memory for devices that only see a part of
flash in the RAM address space, then bit-field NVMCTRL_CTRLB.FLMAP has to be
in sync with the value of symbol __flmap.  That symbol determines the load
memory address (LMA) for rodata when rodata is located in program memory
according to Binutils PR31124.  NVMCTRL_CTRLB.FLMAP determines which 32k block
of program memory is visible in the RAM address space at 0x8000.

This applies to AVR-Dx and AVR-Ex devices like AVR64* (from avrxmega2)
and AVR128* (from avrxmega4).

The new module contains naked .init code that is referenced by the compiler
as __do_flmap_init in order to drag it into the startup code, provided:

o The code is not linked with -mrodata-in-ram.

o The compiler implements PR112944 and is configured with Binutils that
  support PR31124.  That's the case for Binutils v2.42 and GCC v14:
  https://gcc.gnu.org/gcc-14/changes.html#avr

The new code communicates with the default linker scripts for emulations
avrxmega2[_flmap] and avrxmega4[_flmap] by means of a bunch of symbols.
For details, see the added code.  Symbols that are of interest for the user:

o __flmap: Specify which 32k block of program memory is visible in the
  RAM address space.  The (hardware) default is to use the last 32k block.

o __RODATA_FLASH_START__: Specify the byte address of the 32k flash block that
  is visible in the RAM address space.  This takes precedence over __flmap.

o __flmap_lock: When set to a non-zero value, set bit NVMCTRL_CTRLB.FLMAPLOCK
  which will protect NVMCTRL_CTRLB.FLMAP from any future changes.

o __do_flmap_init: Referencing this symbol will drag the part of the startup
  code from flmap_init.S. This can be used when PR31124 is not available but
  the code from flmap_init.S is still wanted, for example by means of linking
  with  -Wl,-u,__do_flmap_init   When the compiler references the code but it
  is *NOT* wanted, then satisfying the symbol by means of linking with, e.g.
  -Wl,--defsym,__do_flmap_init=0  will keep the linker from dragging the code.

The new code will only change NVMCTRL_CTRLB when an avrxmega*_flmap emulation
is used (-mno-rodata-in-ram).  When an avrxmega2/4 emulation (-mrodata-in-ram)
is used, then a conditional RJMP (conditional on the emulation) will bypass
the initialization of NVMCTRL_CTRLB.

	* libc/misc/flmap_init.S: New file.
	* libc/misc/Files.am (dev_asm_sources): Add it.