Issue #563: Support pgm_read macros for Reduced Tiny.
This adds support for Reduced Tiny for some of the features provided
by avr/pgmspace.h.  While not strictly needed, providing them makes
it easier to write cross-target code.

Attribute __progmem__ on Reduced Tiny works different than for
all the other devices:  When taking the address of a symbol that's
attributed as progmem, then the compiler adds an offset of 0x4000
to the value of the symbol.  This means that accessing data in
progmem can be performed by vanilla C/C++ code.  */

Functions like memcpy_P are provided as a weak alias of their
non-_P version.  The __LPM* macros just use open-coded C/C++.

	* common/asmdef.h (TINY_WEAK_ALIAS): New .macro.
	* common/macros.inc: Same.
	* include/avr/pgmspace.h (__LPM, __LPM_word, __LPM_dword)
	(__LPM_float, __strlen_P) [AVR_TINY]: Define.
	* libc/string/memcmp.S (memcmp_P) [AVR_TINY]: Define as .weak
	alias of memcmp.
	* libc/string/memcpy.S: Similar.
	* libc/string/memccpy.S: Similar.
	* libc/string/memchr.S: Similar.
	* libc/string/memrchr.S: Similar.
	* libc/string/strcat.S: Similar.
	* libc/string/strchr.S: Similar.
	* libc/string/strcmp.S: Similar.
	* libc/string/strcpy.S: Similar.
	* libc/string/strlen.S: Similar.
	* libc/string/strstr.S: Similar.
	* libc/string/strncat.S: Similar.
	* libc/string/strncmp.S: Similar.
	* libc/string/strncpy.S: Similar.
	* libc/string/strnlen.S: Similar.
	* libc/string/strrchr.S: Similar.
	* libc/string/strchrnul.S: Similar.
	* libc/string/strcasecmp.S: Similar.
	* libc/string/strncasecmp.S: Similar.
	* libc/pmstring/memccpy_P.S: Wrap code in #ifndef __AVR_TINY__.