iox128a1u.h: Update ADC bit definitions and channel structs
Some bit names disagreed with their counterparts in the datasheet. In other cases, enums were cut short causing entire modes of operation to be completely omitted.

The ADC_CH_t struct was missing the SCAN byte, which led the entire ADC_t struct to be misaligned with the actual locations of latter registers in memory.