vectortable.dox - Rebuild with new generator script.
The script that was used to generate doc/api/vectortable.dox
to date was  $source/xml/patch-header.py.  It is outdated because
the XML format changed (a lot).

This patch adds a new script that distills the vector names and their
comments / documentation from the device header as of #include <avr/io.h>.

This requires a compiler that can pre-process for all supported -mmcu's.

The extracted information is then output in a format / syntax appropriate
for vectortable.dox.  Due to the sheer size of the table with ~650 rows
and ~12000 MCU mentions, the table is currently only included in the
HTML version of the documentation using \htmlonly.

To build vectortable.dox, run in $source/doc/api the shell script

  ./gen-vectortable-dox.sh

Prerequisites:

* Python3

* The used compiler is avr-gcc.  It can be set with CC=

* The list of supported devices is taken from $source/tmp-device-info
  which is generated by bootstrap.  Notice that tmp-device-info is
  NOT part of the source distribution.

doc/api/
	* gen-vectortable-dox.sh: New Shell script, calling...
	* isrs-to-dox.py: ...this new Shell script,
	* isrs-to-dox.py: ...and this new Python3 script.
	* vect_to_desc.py: New Python module.
	* sections.dox: Rebuild.