free() - Avoid traversing freelist twice (#991).
Freelist in free() is walked twice: first time to find the position of
the new chunk, and the second time to see if there is a new top-most
chunk that can be absorbed into __brkval.  The second walk through the
freelist is redundant and introduces visible overhead (benchmarking is
below) when the size of the freelist grows large.

The free() algorithm can be modified to avoid traversing through the entire
list two times.  Most of the code is reused, but the logic is rearranged
in the following way:

1. Check if __brkval can be reduced to fpnew.
2. If 1., then traverse the list to find the last chunk, and see if it
   can be merged with new __brkval.  Merge and eliminate the freelist if
   necessary.  Anyway, we are done. Return.
3. Else (not 1.), transverse the list to find the position of the new chunk.
   Merge the new chunk either at the start, or in the middle, or at the end
   of the freelist.  The list cannot be entirely eliminated in this branch.
   Return.

In this formulation, the list is either traversed once in 2. or in 3.,
but not twice.  This approach shows performance gains in tests such as
malloc-8.c and realloc-3.c where there is a large number of pseudo-random
allocations/deallocations.

Benchmarking results using simulavr with gcc-14.2.0 and -Os optimization
flag are as follows (modified algorithm vs original algorithm, in CPU cycles):

malloc-8.c N=500
2008375 <- modified
2256447 <- original

malloc-8.c N=1000
4660592
5443178

malloc-8.c N=2000
8495885
9677264

malloc-8.c N=5000 (RAND_START_VALUE=1000)
23163097
26296019

realloc-3.c N=2000
2339246
2371133

realloc-3.c N=4000
5441688
5587853

Short tests show little or zero performance gains.