Merge libm/fplib/asmdef.h into common/asmdef.h (#992)
The content of the private header libm/fplib/asmdef.h is mostly
identical to common/asmdef.h.  The main difference is the definition
of macro ENTRY_FLOAT in the former that is missing from the latter.
There is also a difference in the prefix for local symbols between
two files. This commit makes a three-way merge between two files
to provide a unified content.

	* common/asmdef.h (ENTRY_FLOAT): New .macro from libm/fplib/asmdef.h.
	* libm/fplib/Files.am (libm_a_extra_dist) <asmdef.h>: Remove.
	* libm/fplib/asmdef.h: Remove file.