merge libm/fplib/asmdef.h into common/asmdef.h
The content of the private header libm/fplib/asmdef.h is mostly
identical to common/asmdef.h. The main difference is the definition
of the ENTRY_FLOAT in the former that is missing from the latter.
There is also a difference in the prefix for local symbols between
two files. This commit makes a three-way merge between two files
to provide a unified content.