gcrt1.S: Make code to init SP optional (#1011).
This moves the code to init SP from gcrt1.S to a new module hosted
by the device lib.  If the code is not wanted, symbol __init_sp can
by satisfies by -Wl,--defsym,__init_sp=0 or by means of

__asm (".global __init_sp\n"
       "__init_sp = 0");

and it will no more be pulled in from libmcu.a.

	* crt1/gcrt1.S: Move code to init SP to...
	* libc/misc/init_sp.S: ...this new asm module.
	* libc/misc/Files.am (dev_asm_sources) Add  init_sp.S.