asinf, acosf: Rework to reduce asinf error from 8 ULPs to 3 ULPs.
For small (in around [0, 0.25]) inputs, the asinf function used an
ad hoc Taylor polynomial which is sub-optimal.  This patch uses a
degree 5 MiniMax polynomial for asinf and acosf arguments in the
range [0, 1/2), and sqrtf + degree 5 MiniMax polynomial for arguments
in the range [1/2, 1].

The error of acosf ramains at 2 ULPs, but the computation is
faster since it requires 2 multiplications and 1 addition less.
With MUL, the WCET improves from 2900 cycles to 2500 cycles.