fix: Add a docs/examples/ folder to fix #1021 for old releases.
This patch cures issue #1021 (404 for asmdemp.c etc.) for the
existing HTML online documentations in docs/ by adding an examples/
folder parallel to the

dosc/avr-libc-user-manual-<version>/

folders.

The obvious drawback is that all versions of the documentation are
using the same examples/ folder, hence there should be a proper fix
for future releases.