fix: doc/examples/Makefile.am: Install examples/ in $(AVR_LIBC_USER_MANUAL) #1021
This patch installs examples in $(DOC_INST_DIR)/$(AVR_LIBC_USER_MANUAL)/examples
so that examples is inside the documentation tree and not aside of it like
it was the case with the previous path of $(DOC_INST_DIR)/examples.

What currently works to build and install the HTML documentation is:

* In $builddir/doc/api: make html; make install-dox-html
* In $builddir: make install

Notice that this requires an already built and installed toolchain, including
an AVR-LibC installation.

In order to make the PDF documentation, Doxygen v1.9.6 is needed and

* In $builddir/doc/api: make pdf; make install-dox-pdf