perf: configure.ac: Allow jump-tables if PR63223 is fixed.
GCC PR63223 implements reading from jump-tables with ELPM, which
fixes the issue that LPM had when -Text= was used:  Jump tables may
be moved out of the scope of LPM.  This is the reason for why libc
is disabling jump tables on ELPM devices in gen-avr-lib-tree.sh.in.

This patch adds CHECK_JUMP_TABLES_ISSUE as a new configure test,
which uses the compiler version as a proxy to determine whether
PR63223 has been fixed (which is for v4.9.2+).  If so, configure
sets FNO_JUMP_TABLES="".