new: configure.ac (CHECK_AVR_DEVICE): Abort if avr/io.h doesn't compile for at least one device.
This patch extends CHECK_AVR_DEVICE to conduct a test that
compiles <avr/io.h> for at least on device, namely the first
device that is recognized to be supported in C.

The purpose it to exit early when the provided compiler cannot
compile avr/io.h.  It prevents strange configure results when the
provided compiler is too old to build the library.