Update for Arduino Nano
This adds the ATmega328P as a possible MCU configuration.

Some pecularities about the Arduino Nano involve that it does not
connect AREF to anything, so the internal AVcc must be used as
reference.  Also, pre-scale the clock from 16 to 1 MHz to match the
configuration that has been used before.