fix #654: Add alternative to -Wl,-u,vfprintf -lprintf_flt et al.
In addition to the classic  -Wl,-u,vfprintf -lprintf_flt  this
patch allows to use  -Wl,--defsym,vfprintf=vfprintf_flt  instead.
Advantage is that when the application doesn't use printf and links
with -Wl,--gc-sections, then no printf code will be linked.

Similar applies to the _min versions and to vfscanf.