fix: Help AVR-SD: Avoid attribute "section" with switch/case (#1055).
On AVR-SD devices, every instruction that's executed must be
followed by a legal opcode.

One situation where this is not the case is when avr-gcc compiles
a switch table for a function with a "section" attribute.  In order
to meet the section attribute, the jump table labels are placed in
.subsection 1, i.e. directly after the function code.

Therefore, remove ATTRUBUTE_CLIB_SECTION from all libc functions that
have switch statements: vfprintf, vfscanf and strftime.  Instead,
these modules are now compiled with -ffunction-sections.