fix #1060: gcrt1.S: Reference main().
Arduino assumes that main() is referenced from the startup code, which
is no more true since calling main() happens in lib<mcu>.a since v.2.3.0.
The fix adds a reference to main() in crt<mcu>.o, though there may still be
problems in LTO builds.  Building the main() module with -ffat-lto-objects
might mitigate the problem, but to date it's unknown whether there is a
better approach.